Release v0.317.0
This commit is the built version of Nuclide suitable for apm and npm.

+++ pkg/nuclide-hg-repository/lib/HgRepositoryProvider.js
@@ -1,33 +1,57 @@
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the license found in the LICENSE file in
- * the root directory of this source tree.
- *
- * @flow strict-local
- * @format
- */
+'use strict';
 
-import type {NuclideUri} from 'nuclide-commons/nuclideUri';
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
 
-import UniversalDisposable from 'nuclide-commons/UniversalDisposable';
-import {trackTiming} from '../../nuclide-analytics';
-import {
-  RemoteDirectory,
-  getHgServiceByNuclideUri,
-} from '../../nuclide-remote-connection';
-import {HgRepositoryClient} from '../../nuclide-hg-repository-client';
-import {getLogger} from 'log4js';
-import {findHgRepository} from '../../nuclide-source-control-helpers';
-import invariant from 'assert';
+var _UniversalDisposable;
 
-const logger = getLogger('nuclide-hg-repository');
+function _load_UniversalDisposable() {
+  return _UniversalDisposable = _interopRequireDefault(require('../../../modules/nuclide-commons/UniversalDisposable'));
+}
 
-type RefCountedRepo = {
-  refCount: number,
-  repo: HgRepositoryClient,
-};
+var _nuclideAnalytics;
+
+function _load_nuclideAnalytics() {
+  return _nuclideAnalytics = require('../../nuclide-analytics');
+}
+
+var _nuclideRemoteConnection;
+
+function _load_nuclideRemoteConnection() {
+  return _nuclideRemoteConnection = require('../../nuclide-remote-connection');
+}
+
+var _nuclideHgRepositoryClient;
+
+function _load_nuclideHgRepositoryClient() {
+  return _nuclideHgRepositoryClient = require('../../nuclide-hg-repository-client');
+}
+
+var _log4js;
+
+function _load_log4js() {
+  return _log4js = require('log4js');
+}
+
+var _nuclideSourceControlHelpers;
+
+function _load_nuclideSourceControlHelpers() {
+  return _nuclideSourceControlHelpers = require('../../nuclide-source-control-helpers');
+}
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+const logger = (0, (_log4js || _load_log4js()).getLogger)('nuclide-hg-repository'); /**
+                                                                                     * Copyright (c) 2015-present, Facebook, Inc.
+                                                                                     * All rights reserved.
+                                                                                     *
+                                                                                     * This source code is licensed under the license found in the LICENSE file in
+                                                                                     * the root directory of this source tree.
+                                                                                     *
+                                                                                     *  strict-local
+                                                                                     * @format
+                                                                                     */
 
 /**
  * @param directory Either a RemoteDirectory or Directory we are interested in.
@@ -41,62 +65,50 @@ type RefCountedRepo = {
  *    repository (i.e. if it's a remote directory, the URI minus the hostname).
  *  If the directory is not part of a Mercurial repository, returns null.
  */
-function getRepositoryDescription(
-  directory: atom$Directory | RemoteDirectory,
-): ?{
-  originURL: ?string,
-  repoPath: string,
-  workingDirectoryPath: NuclideUri,
-} {
-  if (directory instanceof RemoteDirectory) {
+function getRepositoryDescription(directory) {
+  if (directory instanceof (_nuclideRemoteConnection || _load_nuclideRemoteConnection()).RemoteDirectory) {
     const repositoryDescription = directory.getHgRepositoryDescription();
-    if (
-      repositoryDescription == null ||
-      repositoryDescription.repoPath == null
-    ) {
+    if (repositoryDescription == null || repositoryDescription.repoPath == null) {
       return null;
     }
     const serverConnection = directory._server;
-    const {repoPath, originURL, workingDirectoryPath} = repositoryDescription;
+    const { repoPath, originURL, workingDirectoryPath } = repositoryDescription;
     // These paths are all relative to the remote fs. We need to turn these into URIs.
     const repoUri = serverConnection.getUriOfRemotePath(repoPath);
-    const workingDirectoryUri = serverConnection.getUriOfRemotePath(
-      workingDirectoryPath,
-    );
+    const workingDirectoryUri = serverConnection.getUriOfRemotePath(workingDirectoryPath);
     return {
       originURL,
       repoPath: repoUri,
-      workingDirectoryPath: workingDirectoryUri,
+      workingDirectoryPath: workingDirectoryUri
     };
   } else {
-    const repositoryDescription = findHgRepository(directory.getPath());
+    const repositoryDescription = (0, (_nuclideSourceControlHelpers || _load_nuclideSourceControlHelpers()).findHgRepository)(directory.getPath());
     if (repositoryDescription == null) {
       return null;
     }
-    const {repoPath, originURL, workingDirectoryPath} = repositoryDescription;
+    const { repoPath, originURL, workingDirectoryPath } = repositoryDescription;
     return {
       originURL,
       repoPath,
-      workingDirectoryPath,
+      workingDirectoryPath
     };
   }
 }
 
-export default class HgRepositoryProvider {
+class HgRepositoryProvider {
+  constructor() {
+    this._activeRepositoryClients = new Map();
+  }
   // Allow having multiple project roots under the same repo while sharing
   // the underlying HgRepositoryClient.
-  _activeRepositoryClients: Map<string, RefCountedRepo> = new Map();
 
-  repositoryForDirectory(
-    directory: atom$Directory | RemoteDirectory,
-  ): Promise<?HgRepositoryClient> {
+
+  repositoryForDirectory(directory) {
     return Promise.resolve(this.repositoryForDirectorySync(directory));
   }
 
-  repositoryForDirectorySync(
-    directory: atom$Directory | RemoteDirectory,
-  ): ?HgRepositoryClient {
-    return trackTiming('hg-repository.repositoryForDirectorySync', () => {
+  repositoryForDirectorySync(directory) {
+    return (0, (_nuclideAnalytics || _load_nuclideAnalytics()).trackTiming)('hg-repository.repositoryForDirectorySync', () => {
       try {
         const repositoryDescription = getRepositoryDescription(directory);
         if (!repositoryDescription) {
@@ -106,7 +118,7 @@ export default class HgRepositoryProvider {
         const {
           originURL,
           repoPath,
-          workingDirectoryPath,
+          workingDirectoryPath
         } = repositoryDescription;
 
         // extend the underlying instance of HgRepositoryClient to prevent
@@ -118,31 +130,34 @@ export default class HgRepositoryProvider {
         if (activeRepoClientInfo != null) {
           activeRepoClientInfo.refCount++;
         } else {
-          const hgService = getHgServiceByNuclideUri(workingDirectoryPath);
-          const activeRepoClient = new HgRepositoryClient(repoPath, hgService, {
+          const hgService = (0, (_nuclideRemoteConnection || _load_nuclideRemoteConnection()).getHgServiceByNuclideUri)(workingDirectoryPath);
+          const activeRepoClient = new (_nuclideHgRepositoryClient || _load_nuclideHgRepositoryClient()).HgRepositoryClient(repoPath, hgService, {
             workingDirectoryPath,
-            originURL,
+            originURL
           });
 
           activeRepoClientInfo = {
             refCount: 1,
-            repo: activeRepoClient,
+            repo: activeRepoClient
           };
           activeRepositoryClients.set(repoPath, activeRepoClientInfo);
         }
 
         let destroyed = false;
 
-        const localDisposables = new UniversalDisposable();
+        const localDisposables = new (_UniversalDisposable || _load_UniversalDisposable()).default();
 
         /* eslint-disable no-inner-declarations */
         function ProjectHgRepositoryClient() {
-          this.getProjectDirectory = function(): NuclideUri {
+          this.getProjectDirectory = function () {
             return directory.getPath();
           };
 
-          this.destroy = function(): void {
-            invariant(activeRepoClientInfo != null);
+          this.destroy = function () {
+            if (!(activeRepoClientInfo != null)) {
+              throw new Error('Invariant violation: "activeRepoClientInfo != null"');
+            }
+
             if (!destroyed && --activeRepoClientInfo.refCount === 0) {
               destroyed = true;
               activeRepoClientInfo.repo.destroy();
@@ -154,12 +169,12 @@ export default class HgRepositoryProvider {
           // Allow consumers to use `onDidDestroy` for the ProjectRepos.
           // `getRootRepo()` can be used to add an onDidDestroy for the base
           // repo
-          this.onDidDestroy = function(callback: () => mixed): IDisposable {
+          this.onDidDestroy = function (callback) {
             localDisposables.add(callback);
             return {
               dispose() {
                 localDisposables.remove(callback);
-              },
+              }
             };
           };
         }
@@ -167,16 +182,12 @@ export default class HgRepositoryProvider {
         ProjectHgRepositoryClient.prototype = activeRepoClientInfo.repo;
 
         // $FlowFixMe: this object has an HgRepositoryClient instance in its prototype chain
-        return ((new ProjectHgRepositoryClient(): any): HgRepositoryClient);
+        return new ProjectHgRepositoryClient();
       } catch (err) {
-        logger.error(
-          'Failed to create an HgRepositoryClient for ',
-          directory.getPath(),
-          ', error: ',
-          err,
-        );
+        logger.error('Failed to create an HgRepositoryClient for ', directory.getPath(), ', error: ', err);
         return null;
       }
     });
   }
 }
+exports.default = HgRepositoryProvider;
